back/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.0.0
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2